Add the first step of the algorithm - creating events.
In addition, I define EventRelation to store the flags described in section 4, as well as the sibling of each event. The original paper uses pointers for the sibling, but we can get around this by storing the ID of the sibling. However we also need to store sibling point to make the EventRelation struct self-contained (we won't have access to all events until we join the edges).