Add subdivide_edges to process events and produce events in the result polygon.
This corresponds to the "first stage of the algorithm" from the paper (sections 3 and 4). This is pretty much the bulk of the algorithm, and where all the details end up biting you (from my own experience). Testing subdivide_edges seems hard and not worth it - we will probably get more benefit from testing the entire algorithm at that point.