Add join_contours to create the final polygon.
This maps to the paper's second stage of the algorithm. EventContourFlags corresponds to Listing 4, and compute_contour corresponds to Listing 3. I've also decided not to test all of this since the order of the events is so coupled to the algorithm working. In theory, we could test compute_contour, but the test cases would only work because of some magic order of the events.