Return the source edges of each edge in the result polygon.
This is an extension of the original algorithm. I want users (particularly my own uses lol) to be able to correspond each result edge to a subject or clip edge. This can be useful if for example each polygon is a room and some edges of the room are doors - we may wish to know which edges of the result polygon should still be doors.